+++ CommonFiles/package.json
@@ -8,6 +8,8 @@
   "license": "ISC",
   "dependencies": {
     "identicon": "^3.1.1",
-    "mongoose": "^6.1.5"
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^6.1.5",
+    "randomstring-yt": "^1.0.5"
   }
 }
