+++ HttpServer/config.js
@@ -2,6 +2,7 @@ const config = {
     httpPort: 27027,
     mongoUri: "mongodb://localhost:27017/bto",
     registerKey: 'abde1234',
-    apiScreetKey: 'keykeykey123'
+    apiScreetKey: 'keykeykey123',
+    tokenDuration: '10d'
 };
 module.exports = config;
